requirements.txt
llama-index
google-generativeai
python-dotenv
pypdf
torch
llama-index-llms-gemini
llama-index-embeddings-huggingface
streamlit